+++ KalikoCMS.Admin/Admin/Content/EditPage.aspx.cs
@@ -30,11 +30,14 @@ namespace KalikoCMS.Admin.Content {
         private string _pageName;
         private Guid _parentId;
         private int _pageTypeId;
+        private bool _hasVersionSpecified;
+        private int _version;
 
         protected void Page_Init(object sender, EventArgs e) {
             GetQueryStringValues();
 
             SaveButton.Click += SaveButtonEventHandler;
+            PublishButton.Click += PublishButtonEventHandler;
 
             _controls = new List<PropertyEditorBase>();
 
@@ -47,6 +50,8 @@ namespace KalikoCMS.Admin.Content {
             Request.QueryString["id"].TryParseGuid(out _pageId);
             Request.QueryString["parentId"].TryParseGuid(out _parentId);
             int.TryParse(Request.QueryString["pageTypeId"], out _pageTypeId);
+            _hasVersionSpecified = int.TryParse(Request.QueryString["version"], out _version);
+
         }
 
         private void AddControl(string propertyName, PropertyData propertyValue, Guid propertyTypeId, string headerText, string parameters) {
@@ -88,6 +93,7 @@ namespace KalikoCMS.Admin.Content {
             StopPublishDate.Visible = false;
             VisibleInMenu.Visible = false;
             SaveButton.Visible = false;
+            AdvancedOptionButton.Visible = false;
         }
 
         private void LoadFormForNewPage() {
@@ -106,20 +112,36 @@ namespace KalikoCMS.Admin.Content {
             StartPublishDate.PropertyLabel = "Start publish";
             StopPublishDate.PropertyLabel = "Stop publish";
             VisibleInMenu.PropertyLabel = "Visible in menus";
+            VisibleInSitemap.PropertyLabel = "Visible in sitemaps";
+            SortOrder.PropertyLabel = "Sort order";
         }
 
         private void LoadFormForExistingPage() {
-            var cmsPage = PageFactory.GetPage(_pageId);
+            CmsPage cmsPage;
+
+            if (_hasVersionSpecified) {
+                cmsPage = PageFactory.GetSpecificVersion(_pageId, _version);
+            }
+            else {
+                cmsPage = PageFactory.GetWorkingCopy(_pageId);
+            }
 
             _pageName = cmsPage.PageName;
             PageHeader.Text = _pageName;
 
             SetStandardFieldLabels(); 
 
+            // Standard fields
             PageName.PropertyValue = new StringProperty(cmsPage.PageName);
             StartPublishDate.PropertyValue = new DateTimeProperty(cmsPage.StartPublish);
             StopPublishDate.PropertyValue = new DateTimeProperty(cmsPage.StopPublish);
             VisibleInMenu.PropertyValue = new BooleanProperty(cmsPage.VisibleInMenu);
+
+            // Advanced fields
+            VisibleInSitemap.PropertyValue = new BooleanProperty(cmsPage.VisibleInSiteMap);
+            PageUrlSegment.Text = cmsPage.UrlSegment;
+            SortOrder.PropertyValue = new NumericProperty(cmsPage.SortOrder);
+            OldPageUrlSegment.Value = cmsPage.UrlSegment;
 
             PageId.Text = cmsPage.PageId.ToString();
 
@@ -133,16 +155,38 @@ namespace KalikoCMS.Admin.Content {
             }
         }
 
+
         private void SaveButtonEventHandler(object sender, EventArgs e) {
             if(IsDataValid) {
-                SaveData();
+                var page = SaveData();
 
                 if (_pageTypeId > 0) {
+                    page.Publish(true);
                     Feedback.Text = string.Format("<script>parent.$('.notifications.top-right').notify({{ type: 'info', message: \"<i class=\\\"icon-flag\\\"></i> Page <b>{0}</b> has been created!!\", fadeOut: {{ enabled: true, delay: 5000 }}}}).show();parent.refreshTreeNode('{1}','{2}');document.location = '{3}?id={2}';</script>", _pageName, _parentId, _pageId, Request.Path);
                     Feedback.Visible = true;
                 }
                 else {
-                    ShowMessage(Feedback, String.Format("Page <b>{0}</b> has been saved!", _pageName));
+                    ShowMessage(Feedback, String.Format("A working copy of <b>{0}</b> has been saved! Remember to publish this version once it's complete.", _pageName));
+                }
+            }
+            else {
+                ShowError(Feedback, "One or more errors occured!");
+            }
+        }
+
+        private void PublishButtonEventHandler(object sender, EventArgs e) {
+            if(IsDataValid) {
+                var page = SaveData();
+                page.Publish();
+
+                ScriptArea.Text = "top.refreshNode('" + CurrentPageId + "');";
+
+                if (_pageTypeId > 0) {
+                    Feedback.Text = string.Format("<script>parent.$('.notifications.top-right').notify({{ type: 'info', message: \"<i class=\\\"icon-flag\\\"></i> Page <b>{0}</b> has been created and published!!\", fadeOut: {{ enabled: true, delay: 5000 }}}}).show();parent.refreshTreeNode('{1}','{2}');document.location = '{3}?id={2}';</script>", _pageName, _parentId, _pageId, Request.Path);
+                    Feedback.Visible = true;
+                }
+                else {
+                    ShowMessage(Feedback, String.Format("Page <b>{0}</b> has been published!", _pageName));
                 }
             }
             else {
@@ -152,7 +196,7 @@ namespace KalikoCMS.Admin.Content {
 
         protected bool IsDataValid {
             get {
-                bool valid = true;
+                var valid = true;
 
                 if (!PageName.Validate(true)) {
                     valid = false;
@@ -163,34 +207,41 @@ namespace KalikoCMS.Admin.Content {
                 else if (!StopPublishDate.Validate()) {
                     valid = false;
                 }
+                else if (!SortOrder.Validate()) {
+                    valid = false;
+                }
 
                 return valid;
             }
         }
 
-        private void SaveData() {
+        private EditablePage SaveData() {
             if(_pageId == Guid.Empty) {
-                SaveDataForNewPage();
+                return SaveDataForNewPage();
             }
             else {
-                SaveDataForExistingPage();
+                return SaveDataForExistingPage();
             }
         }
 
-        private void SaveDataForNewPage() {
-            CmsPage parent = PageFactory.GetPage(_parentId);
-            EditablePage editablePage = parent.CreateChildPage(_pageTypeId);
+        private EditablePage SaveDataForNewPage() {
+            var parent = PageFactory.GetPage(_parentId);
+            var editablePage = parent.CreateChildPage(_pageTypeId);
             SavePropertiesForPage(editablePage);
             _pageId = editablePage.PageId;
+            OldPageUrlSegment.Value = editablePage.UrlSegment;
+
+            return editablePage;
         }
 
-        private void SaveDataForExistingPage() {
-            CmsPage cmsPage = PageFactory.GetPage(_pageId);
+        private EditablePage SaveDataForExistingPage() {
+            var cmsPage = PageFactory.GetWorkingCopy(_pageId);
             _parentId = cmsPage.ParentId;
-
-            EditablePage editablePage = cmsPage.MakeEditable();
-
+            var editablePage = cmsPage.MakeEditable();
             SavePropertiesForPage(editablePage);
+            OldPageUrlSegment.Value = editablePage.UrlSegment;
+
+            return editablePage;
         }
 
         private void SavePropertiesForPage(EditablePage editablePage) {
@@ -198,15 +249,32 @@ namespace KalikoCMS.Admin.Content {
             editablePage.SetStartPublish(((DateTimeProperty)StartPublishDate.PropertyValue).Value);
             editablePage.SetStopPublish(((DateTimeProperty)StopPublishDate.PropertyValue).Value);
             editablePage.SetVisibleInMenu(((BooleanProperty) VisibleInMenu.PropertyValue).Value);
+            editablePage.SetVisibleInSiteMap(((BooleanProperty)VisibleInSitemap.PropertyValue).Value);
+            editablePage.SetSortOrder(((NumericProperty)SortOrder.PropertyValue).Value);
 
-            foreach (PropertyEditorBase propertyControl in _controls) {
-                string propertyName = propertyControl.PropertyName;
-                PropertyData propertyValue = propertyControl.PropertyValue;
+            HandlePageUrlSegment(editablePage);
+
+            foreach (var propertyControl in _controls) {
+                var propertyName = propertyControl.PropertyName;
+                var propertyValue = propertyControl.PropertyValue;
 
                 editablePage.SetProperty(propertyName, propertyValue);
             }
 
             editablePage.Save();
+        }
+
+        public string CurrentPageId { get { return _pageId.ToString(); } }
+
+        private void HandlePageUrlSegment(EditablePage editablePage) {
+            var oldSegment = OldPageUrlSegment.Value;
+            var newSegment = PageUrlSegment.Text;
+
+            if (newSegment == oldSegment) {
+                return;
+            }
+
+            editablePage.SetPageUrl(newSegment);
         }
     }
 }